requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.4
psycopg2-binary==2.9.6
requests==2.31.0
python-dotenv==1.0.0
linkedin-api==2.0.0